Extracted BidsDataFrame class outside of loader/utils.py
