+++ testing/unit_tests/test_automate_training.py
@@ -87,6 +87,8 @@ For ``mult_params``:
 """
 
 import pytest
+
+from ivadomed.loader.bids_dataframe import BidsDataframe
 from ivadomed.scripts.automate_training import make_config_list, get_param_list, \
     HyperparameterOption
 from ivadomed.loader import utils as imed_loader_utils
@@ -717,7 +719,7 @@ def test_config_sha256(download_data_testing_test_files, initial_config):
         "multichannel": False
     }
 
-    bids_df = imed_loader_utils.BidsDataframe(loader_params, __tmp_dir__, derivatives=True)
+    bids_df = BidsDataframe(loader_params, __tmp_dir__, derivatives=True)
     generate_sha_256(initial_config, bids_df.df, file_lst)
     assert(initial_config['training_sha256']['sub-unf01_T2w.nii.gz'] ==
            'f020b368fea15399fa112badd28b2df69e044dba5d23b3fe1646d12d7d3d39ac')
