+++ testing/unit_tests/test_bounding_box.py
@@ -4,6 +4,7 @@ import os
 import json
 import shutil
 
+from ivadomed.loader.bids_dataframe import BidsDataframe
 from ivadomed.loader import loader as imed_loader
 from ivadomed.loader import utils as imed_loader_utils
 from ivadomed.object_detection import utils as imed_obj_detect
@@ -93,7 +94,7 @@ def test_bounding_box(download_data_testing_test_files, train_lst, target_lst, c
     # Update loader_params with config
     loader_params.update(config)
 
-    bids_df = imed_loader_utils.BidsDataframe(loader_params, __tmp_dir__, derivatives=True)
+    bids_df = BidsDataframe(loader_params, __tmp_dir__, derivatives=True)
 
     ds = imed_loader.load_dataset(bids_df, **loader_params)
