+++ testing/unit_tests/test_sampler.py
@@ -2,6 +2,8 @@ import numpy as np
 import pytest
 import torch.backends.cudnn as cudnn
 from torch.utils.data import DataLoader
+
+from ivadomed.loader.bids_dataframe import BidsDataframe
 from ivadomed import utils as imed_utils
 from ivadomed.loader import utils as imed_loader_utils, loader as imed_loader
 from testing.unit_tests.t_utils import create_tmp_dir,  __data_testing_dir__, __tmp_dir__, download_data_testing_test_files
@@ -71,7 +73,7 @@ def test_sampler(download_data_testing_test_files, transforms_dict, train_lst, t
         "multichannel": False
     }
     # Get Training dataset
-    bids_df = imed_loader_utils.BidsDataframe(loader_params, __tmp_dir__, derivatives=True)
+    bids_df = BidsDataframe(loader_params, __tmp_dir__, derivatives=True)
     ds_train = imed_loader.load_dataset(bids_df, **loader_params)
 
     print('\nLoading without sampling')
